fix: update lib generator to create new console file
